+++ projects/ngx-sub-form/src/public_api.ts
@@ -10,5 +10,5 @@ export * from './lib/shared/ngx-sub-form-utils';
 export * from './lib/shared/ngx-sub-form.types';
 
 export * from './lib/helpers';
-export * from './lib/ngx-sub-form';
+export * from './lib/create-form';
 export * from './lib/ngx-sub-form.types';
